Undo go:build statements, make code a little DRY
